+++ src/ralph/models/edx/__init__.py
@@ -4,6 +4,13 @@
 
 from ralph.models.edx.navigational import statements
 
+from .drag_and_drop.statements import (
+    EdxDragAndDropV2FeedbackClosed,
+    EdxDragAndDropV2FeedbackOpened,
+    EdxDragAndDropV2ItemDropped,
+    EdxDragAndDropV2ItemPickedUp,
+    EdxDragAndDropV2Loaded,
+)
 from .enrollment.statements import (
     EdxCourseEnrollmentActivated,
     EdxCourseEnrollmentDeactivated,